scripts/elevation-solver/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
numpy==2.2.5
    # via -r requirements.in
pyyaml==6.0.2
    # via -r requirements.in
